+++ test/typescript/apis/project/cds-services.ts
@@ -24,10 +24,6 @@ cds.connect({kind: 'odata', model:'some/imported/model', service: 'BusinessPartn
 // basic properties
 srv.name.length
 srv.entities[0] = Books // same type
-srv.entities('namespace')
-srv.events('namespace')
-srv.types('namespace')
-srv.operations('namespace')
 
 await srv.init()
 
@@ -463,21 +459,28 @@ await cds.db.run ( SELECT.from(Books) )
 await cds.tx (async (tx) => {
   await tx.run(SELECT(1).from(Books,201).forUpdate())
 })
-cds.db.entities('draftModelAuth')
+cds.entities('draftModelAuth')
 
 //tests outbox
 const outboxedService = cds.outboxed(srv)
 await outboxedService.send({ event: 'feeEstimation', entity: networkGroups, data: {name:'Volta'}})
 await cds.unboxed(outboxedService).send({ event: 'feeEstimation', entity: networkGroups, data: {name:'Volta'}})
 
-srv.entities('namespace');
-[...srv.entities('namespace')].map(e => e.keys); // .keys only available on entities
+srv.entities;
+[...srv.entities].map(e => e.keys); // .keys only available on entities
 // @ts-expect-error
-[...srv.events('namespace')].map(e => e.keys);
-[...srv.events('namespace')].map(e => e.elements)
+[...srv.events].map(e => e.keys);
+[...srv.events].map(e => e.elements)
+
+// @ts-expect-deprecation
+srv.entities('namespace');
+srv.events('namespace');
+srv.types('namespace');
+srv.operations('namespace');
+srv.actions('namespace');
 
 // @ts-expect-error
-srv.entities('namespace')('and again')
+cds.entities('namespace')('and again')
 
 type ActionType = HandlerFunction<typeof unboundAction>
 srv.on(unboundAction, externalActionHandler)